message view tests mostly done

+++ app.py
@@ -323,12 +323,15 @@ def messages_destroy(message_id):
     if not g.user:
         flash("Access unauthorized.", "danger")
         return redirect("/")
-
+    
     msg = Message.query.get(message_id)
-    db.session.delete(msg)
-    db.session.commit()
-
-    return redirect(f"/users/{g.user.id}")
+    if msg.user_id == g.user.id:
+        db.session.delete(msg)
+        db.session.commit()
+        return redirect(f"/users/{g.user.id}")
+    else:
+        flash("Delete Your Own Damn Messages!")
+        return redirect(f"/users/{g.user.id}")
 
 
 ##############################################################################
